[SER-153] Data migration. UI fix. Make export path non mandatory field in data migration [SER-147] Move redirect functionality to core
