+++ plugins/data_migration/api/api.js
@@ -955,72 +955,6 @@ function trim_ending_slashes(address) {
         return true;
     });
 
-    //for redirect
-    plugins.register("/sdk", function(ob) {
-        var params = ob.params,
-            app = ob.app;
-        if (!params.cancelRequest && app.redirect_url && app.redirect_url !== '') {
-            var newPath = params.urlParts.path;
-
-            //check if we have query part
-            if (newPath.indexOf('?') === -1) {
-                newPath += "?";
-            }
-
-            var opts = {
-                uri: app.redirect_url + newPath + '&ip_address=' + params.ip_address,
-                method: 'GET'
-            };
-
-            //should we send post request
-            if (params.req.method.toLowerCase() === 'post') {
-                opts.method = "POST";
-                //check if we have body from post method
-                if (params.req.body) {
-                    opts.json = true;
-                    opts.body = params.req.body;
-                }
-            }
-
-            request(opts, function(error, response, body) {
-
-                var code = 400;
-                var message = "Redirect error. Tried to redirect to:" + app.redirect_url;
-
-                if (response && response.statusCode) {
-                    code = response.statusCode;
-                }
-
-
-                if (response && response.body) {
-                    try {
-                        var resp = JSON.parse(response.body);
-                        message = resp.result || resp;
-                    }
-                    catch (e) {
-                        if (response.result) {
-                            message = response.result;
-                        }
-                        else {
-                            message = response.body;
-                        }
-                    }
-                }
-                if (error) { //error
-                    console.log("Redirect error", error, body, opts, app, params);
-                }
-
-
-                if (plugins.getConfig("api", params.app && params.app.plugins, true).safe) {
-                    common.returnMessage(params, code, message);
-                }
-            });
-            params.cancelRequest = "Redirected: " + app.redirect_url;
-            params.waitForResponse = true;
-            return false;
-        }
-        return false;
-    });
 }(pluginOb));
 
 module.exports = pluginOb;
