+++ plugins/data_migration/frontend/public/templates/drawer-export.html
@@ -39,7 +39,7 @@
               </span>
               <cly-tooltip-icon icon="ion-help-circled"></cly-tooltip-icon>
             </div>
-            <validation-provider name="target_path" rules="required" v-slot="v">
+            <validation-provider name="target_path" rules="" v-slot="v">
               <el-input
                 :placeholder="i18n('data-migration.enter-your-export-folder')"
                 v-model="drawerScope.editedObject.target_path">
